removing .env and testing
