+++ src/chapters/workshop3/rds/rds-setup.js
@@ -142,18 +142,11 @@ python_version = "*"
 
 **Security Note**: We use \`os.getenv()\` instead of hardcoding values because database passwords should never be stored directly in code files.
 
-Edit \`rockproject/settings.py\` to support PostgreSQL and environment variables:
+Edit \`rockproject/settings.py\` to support PostgreSQL:
 **Add these imports at the top:**
 \`\`\`python
 from pathlib import Path
 import os
-from dotenv import load_dotenv
-
-# Load environment variables from .env file
-# Only load .env in development
-ENV_PATH = Path(__file__).resolve().parent.parent / '.env'
-if ENV_PATH.exists():
-    load_dotenv(dotenv_path=ENV_PATH)
 
 \`\`\`
 Scroll down and replace \`DATABASES\` with:
@@ -183,68 +176,22 @@ DATABASES = {
 - Share code safely without exposing passwords
 
 **How it works**:
-1. \`python-dotenv\` reads your \`.env\` file when Django starts
-2. \`os.getenv('DB_PASSWORD')\` retrieves the password from environment variables
-3. Django uses these values to connect to the database
+1. Save environment variables in in github secrets (Coming up in following steps).
+2. Docker run commands will reference these to set environment variables on the container
+3. \`os.getenv('DB_PASSWORD')\` retrieves the password from environment variables
+4. Django uses these values to connect to the database
 
-**Real-world analogy**: It's like having a locked box (\`.env\` file) for your house key (database password), instead of leaving the key under a rock (in your code) where anyone can find it.
-
-### Create Environment Variables Template
-Create a \`.env.example\` file in your repository root:
-
-\`\`\`bash
-# Database Configuration
-DB_NAME=rockofages
-DB_USER=rockadmin
-DB_PASSWORD=your-secure-password-here
-DB_HOST=rock-of-ages-db.c9xkv8example.us-east-2.rds.amazonaws.com
-DB_PORT=5432
-\`\`\`
-
-### Create Your Local Environment File
-\`\`\`bash
-# Copy the example to create your actual .env file
-cp .env.example .env
-\`\`\`
-
-Edit the \`.env\` file with your actual RDS information:
-\`\`\`bash
-DB_NAME=rockofages
-DB_USER=rockadmin
-DB_PASSWORD=YourActualRDSPassword123!  # ← Use your actual password
-DB_HOST=rock-of-ages-db.c9xkv8actual.us-east-2.rds.amazonaws.com  # ← Use your actual endpoint
-DB_PORT=5432
-\`\`\`
 
 ### Update Your Database Seed Script
 
 **Understanding the seed script:**
 This script automates the process of setting up your database tables and loading sample data. Think of it like moving into a new house - you need to build the rooms (create tables) before you can put furniture in them (load data).
 
-**Why this order matters:**
-\`\`\`
-Environment Variables → Database Connection → Table Structure → Sample Data
-\`\`\`
 
-Each step depends on the previous one:
-- Can't connect without credentials
-- Can't add data without tables
-- Can't create app tables without Django's foundation tables
-
-Edit \`seed_database.sh\` to load environment variables:
+Replace \`seed_database.sh\` to remove references to sqllite:
 
 \`\`\`bash
 #!/bin/bash
-
-# Load environment variables from .env if it exists
-if [ -f .env ]; then
-  echo "📦 Loading environment from .env..."
-  set -a
-  source .env
-  set +a
-else
-  echo "🚨 .env not found. Assuming environment variables are set externally (e.g., via Docker -e flags)"
-fi
 
 echo "🗄️  Setting up PostgreSQL database..."
 
@@ -270,184 +217,88 @@ python3 manage.py loaddata rocks
 echo "✅ Database setup complete!"
 \`\`\`
 
-**How each step works:**
+### Update tests
+In \`rockapi/test.py\` replace the tests with 
 
-**1. Loading Environment Variables (\`source .env\`(local) Docker -e flags(production))**
-- Makes your database connection info available to Django
-- Like giving Django the keys to connect to your database
+\`\`\`python
+from django.test import SimpleTestCase
+from rest_framework.test import APIClient
+from rest_framework import status
 
-**2. Running Migrations (\`python3 manage.py migrate\`)**
-- Creates Django's standard tables (user accounts, sessions, etc.)
-- Like building the foundation of a house
 
-**3. Creating App Migrations (\`makemigrations rockapi\`)**
-- Looks at your models (Rock.py, Type.py) and creates instructions for building tables
-- Like creating blueprints for your custom rooms
+class SanityTests(SimpleTestCase):
+    def setUp(self):
+        self.client = APIClient()
 
-**4. Applying App Migrations (\`migrate rockapi\`)**
-- Actually creates your app's tables in the database
-- Like building those custom rooms from the blueprints
+    def test_math_still_works(self):
+        self.assertEqual(2 + 2, 4)
 
-**5. Loading Fixtures (\`loaddata\`)**
-- Puts sample data into your tables
-- Like furnishing the rooms with sample furniture
----
+    def test_uppercase(self):
+        self.assertEqual("rock".upper(), "ROCK")
 
-## Step 3: Test Locally with Docker
-
-Before deploying, test your changes locally using the same containerized approach you'll use in production.
-
-### Build the Updated Container
-\`\`\`bash
-# Build the container with PostgreSQL support
-docker build -t rock-of-ages-api .
+    def test_api_mock(self):
+        # This won't hit a real view, but it shows test usage
+        response = self.client.get('/fake-url')
+        self.assertIn(response.status_code, [status.HTTP_404_NOT_FOUND, status.HTTP_200_OK])
 \`\`\`
 
-### Run Container with Database Connection
-\`\`\`bash
-# Run with environment file (container handles all dependencies)
-docker run --env-file .env -p 8000:8000 rock-of-ages-api
+This is only getting updated because the old tests relied on the sqlite database. FYI, these aren't real tests you will see on real projects its just to keep the test step working in github actions without adding extra complexity.  
+
+
+### Update GitHub Actions
+
+Replace \`.github/workflows/deploy.yml\` with:
+
+\`\`\`yaml
+name: Deploy to EC2
+
+on:
+  workflow_dispatch:  # Manual only
+
+permissions:
+  id-token: write
+  contents: read
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: aws-actions/configure-aws-credentials@v3
+        with:
+          role-to-assume: \${{ vars.OIDC_ROLE_TO_ASSUME }}
+          aws-region: \${{ vars.AWS_REGION }}
+
+      - uses: aws-actions/amazon-ecr-login@v2
+
+      - name: Trigger remote deployment on EC2 via SSM
+        run: |
+          aws ssm send-command \\
+            --instance-ids "\${{ vars.EC2_INSTANCE_ID }}" \\
+            --document-name "AWS-RunShellScript" \\
+            --comment "Manual deploy from GitHub Actions" \\
+            --parameters commands='[
+              "IMAGE=\\"\${{ vars.ECR_REGISTRY }}/\${{ vars.ECR_REPOSITORY }}:latest\\"",
+              "aws ecr get-login-password --region \${{ vars.AWS_REGION }} | docker login --username AWS --password-stdin \${{ vars.ECR_REGISTRY }}",
+              "docker pull \\"$IMAGE\\"",
+              "docker stop rock-of-ages-api || true",
+              "docker rm rock-of-ages-api || true",
+              "docker run --pull always -d --name rock-of-ages-api -p 80:8000 -e DB_NAME=\${{ secrets.DB_NAME }} -e DB_USER=\${{ secrets.DB_USER }} -e DB_PASSWORD=\${{ secrets.DB_PASSWORD }} -e DB_HOST=\${{ secrets.DB_HOST }} -e DB_PORT=\${{ secrets.DB_PORT }} \\"$IMAGE\\""
+            ]' \\
+            --region \${{ vars.AWS_REGION }}
 \`\`\`
 
-**Expected output:**
-\`\`\`
-🗄️ Setting up PostgreSQL database...
-⚙️ Running Django migrations...
-📋 Creating app-specific migrations...
-🔧 Applying app migrations...
-📊 Loading seed data...
-✅ Database setup complete!
-System check identified no issues (0 silenced).
-Django version 5.2.3, using settings 'rockproject.settings'
-Starting development server at http://0.0.0.0:8000/
-Quit the server with CONTROL-C.
-\`\`\`
+#### What’s happening here?
+We are running the same docker commands in our ec2 instance but this time with our database environment variables using the -e flag. 
 
-**✅ If you see this output, everything worked!** The container:
-1. Connected to your RDS database
-2. Created all necessary tables
-3. Loaded sample data (users, rock types, rocks)
-4. Started the API server
----
-
-### Test Database Connection Directly
-
-Before testing through the API, let's verify the database setup by connecting directly to your RDS instance. This helps you understand what's happening "under the hood" and gives you valuable database inspection skills.
-
-**Why test the database directly?**
-- **Debugging**: When APIs fail, you can check if the problem is in the database or application layer
-- **Data Verification**: Confirm that migrations and seed data loaded correctly
-- **Professional Skill**: Database inspection is a common developer task
-- **Understanding**: See the actual tables and data your API is working with
-
-#### Install PostgreSQL Explorer Extension
-
-1. **Open VS Code**
-2. **Go to Extensions** (Ctrl+Shift+X or Cmd+Shift+X)
-3. **Search for** "PostgreSQL" by Chris Kolkman
-4. **Install** the PostgreSQL extension
-
-#### Connect to Your RDS Database
-
-1. **Open VS Code Command Palette** (Ctrl+Shift+P or Cmd+Shift+P)
-2. **Type**: "PostgreSQL: New Connection"
-3. **Fill in connection details**:
-   - **Hostname**: Your RDS endpoint (from your .env file)
-   - **User**: \`rockadmin\`
-   - **Password**: Your RDS password
-   - **Port**: \`5432\`
-   - **Database**: \`rockofages\`
-   - **SSL**: \`require\`
-
-#### Explore Your Database
-
-Once connected, you should see your database in the PostgreSQL Explorer panel. Expand the sections to see:
-
-**📁 Schemas → public → Tables**
-- \`auth_user\` (Django users)
-- \`authtoken_token\` (API authentication tokens)
-- \`rockapi_type\` (Rock types: Igneous, Metamorphic, etc.)
-- \`rockapi_rock\` (Individual rocks with owners)
-
-#### Run Test Queries
-
-Right-click on your database connection and select "New Query" to run these verification queries:
-
-**1. Check all rock types:**
-\`\`\`sql
-SELECT * FROM rockapi_type;
-\`\`\`
-**Expected result:** 5 rock types (Metamorphic, Igneous, Sedimentary, Shale, Basalt)
-
-**2. Check all rocks with their types and owners:**
-\`\`\`sql
-SELECT 
-    r.name as rock_name,
-    r.weight,
-    t.label as rock_type,
-    u.first_name,
-    u.last_name
-FROM rockapi_rock r
-JOIN rockapi_type t ON r.type_id = t.id
-JOIN auth_user u ON r.user_id = u.id;
-\`\`\`
-**Expected result:** 3 rocks owned by John Doe and Jane Smith
-
-**3. Count total records:**
-\`\`\`sql
-SELECT 
-    (SELECT COUNT(*) FROM rockapi_rock) as total_rocks,
-    (SELECT COUNT(*) FROM rockapi_type) as total_types,
-    (SELECT COUNT(*) FROM auth_user) as total_users;
-\`\`\`
-**Expected result:** 3 rocks, 5 types, 2 users
-
-**✅ If you see the expected data, your database setup is perfect!**
-
-**Understanding what you're seeing:**
-- **Django tables**: Tables starting with \`auth_\` are Django's built-in user management
-- **Your app tables**: Tables starting with \`rockapi_\` are from your Rock of Ages models
-- **Relationships**: The JOIN queries show how your data connects across tables
-- **Data integrity**: All foreign key relationships are working correctly
-
-### Test with Yaak
-
-Now test your API to ensure it's correctly reading from the database:
-
-**Test: Register a New User**
-1. **Method**: POST
-2. **URL**: \`http://localhost:8000/register\`
-3. **Headers**: Content-Type: \`application/json\`
-4. **Body** (JSON):
-   \`\`\`json
-   {
-     "email": "test@example.com",
-     "password": "testpass123",
-     "first_name": "Test",
-     "last_name": "User"
-   }
-   \`\`\`
-
-**Expected Response:**
-\`\`\`json
-{
-  "token": "abc123def456..."
-}
-\`\`\`
-
-
-**💡 Database verification**: After registering, you can run this query in VS Code to see your new user:
-\`\`\`sql
-SELECT * FROM auth_user ORDER BY date_joined DESC LIMIT 1;
-\`\`\`
 
 
 ## What You’ve Accomplished
 
 - ✅ Created RDS PostgreSQL DB
 - ✅ Updated Django API for PostgreSQL
-- ✅ Secured credentials using .env and GitHub Secrets
-- ✅ Verified functionality locally
